chore(refactor) use hex architecture and golang standard project layout
