+++ CHANGELOG.md
@@ -1,0 +1,39 @@
+cl-# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog],
+and this project adheres to [Semantic Versioning].
+
+## [Unreleased]
+
+- /
+
+## [0.0.2] - 2023-03-04
+
+### Added
+
+### Changed
+
+- Refactored to hexagonal architecture and structuring with https://github.com/golang-standards/project-layout
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+## [0.0.1] - 2023-03-04
+
+- initial release
+
+<!-- Links -->
+[keep a changelog]: https://keepachangelog.com/en/1.0.0/
+[semantic versioning]: https://semver.org/spec/v2.0.0.html
+
+<!-- Versions -->
+[unreleased]: https://github.com/simonireilly/go-gitignore-it/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/simonireilly/go-gitignore-it/compare/v0.0.1...v0.0.2
+[0.0.1]: https://github.com/simonireilly/go-gitignore-it/releases/tag/v0.0.1
