+++ cmd/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/simonireilly/go-gitignore-it/internal/app"
+
+func main() {
+	app.Run()
+}
