+++ internal/app/run.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"fmt"
@@ -8,6 +8,8 @@ import (
 
 	"github.com/gobuffalo/packr"
 	"github.com/ktr0731/go-fuzzyfinder"
+
+	"github.com/simonireilly/go-gitignore-it/internal/ports"
 )
 
 // gitignore - A struct that holds names and metadata for gitignore files
@@ -16,14 +18,14 @@ type gitignore struct {
 	content string
 }
 
-func main() {
+func Run() {
 	// packr box holds byte strings for gitignores
 	box := packr.NewBox("./gitignore")
 
 	// invoke a fuzzy search
 	gi := getGitignoreFiles(box)
 
-	idx := searchGitignores(gi)
+	idx := searchGitIgnores(gi)
 
 	c := []byte(gi[idx[0]].content)
 
@@ -43,7 +45,7 @@ func main() {
 
 // getGitignoreFiles needs to pack the binaries for the gitignores, fetch them from the box
 // then return them as a slice of []gitignore structs
-func getGitignoreFiles(b packr.Box) []gitignore {
+func getGitignoreFiles(b ports.FilesPort) []gitignore {
 	var g []gitignore
 	list := b.List()
 
@@ -58,7 +60,7 @@ func getGitignoreFiles(b packr.Box) []gitignore {
 	return g
 }
 
-func searchGitignores(g []gitignore) []int {
+func searchGitIgnores(g []gitignore) []int {
 	idx, err := fuzzyfinder.FindMulti(
 		g,
 		func(i int) string {
